void commitFixAutomaticallyForTCA2CrashOnTooLongFilenames();
the outputFilenamePathSuggestionString for the .xml is the concatenation
of the input file names, which in some cases ends up being >255 bytes
and giving a crash with a GUI alert(!) wider than my monitor. Added
option -cli-plain that only does the plain .txt output, since we don't
use the .xml. Should be a bit safer.

Also, parallellize now ensures enough leeway on the .txt filenames
suggested by TCA2.

git-svn-id: https://victorio.uit.no/langtech/trunk/tools/CorpusTools@114653 c7155fb1-f0a7-4240-a2fc-2600b6f42f90